+++ RELEASE_NOTES.md
@@ -1,4 +1,8 @@
 
+#### 0.1.5-beta - February 24 2017
+* Implement SOQL Count().
+* Fix: assign entity Id after insert operation.
+
 #### 0.1.4-beta - February 23 2017
 * Fix members select issues.
 * Fix anonymous types select issues.